Feat : 사용자 목표 추가 (#25)

+++ src/main/java/com/ripple/BE/user/dto/response/QuestResponse.java
@@ -2,7 +2,7 @@ package com.ripple.BE.user.dto.response;
 
 import com.ripple.BE.user.dto.QuestDTO;
 
-public record QuestResponse(Long conceptProgress, Long quizProgress, Long articleProgress) {
+public record QuestResponse(int conceptProgress, int quizProgress, int articleProgress) {
 
     public static QuestResponse toQuestResponse(QuestDTO questDTO) {
         return new QuestResponse(
